requirements.txt
asgiref==3.2.10
Django==3.1
django-tinymce==3.2.0
pytz==2020.1
sqlparse==0.3.1
